separate logic from app.py
